Copied and updated the comments in header file
I was using an older version of the mcp_can library and had put a lot of comments into it to help me.  I had also exposed more of the error handling.  The latest version of the library has also exposed the errors so I don't need to do that any more but I like having loads of comments against the public methods.  Note this is not backwards compatible with older SW so all CAN routines will need updating.